Split OpenAI usage from eval history
