+++ apps/backoffice/src/components/backoffice-layout/types.ts
@@ -2,6 +2,7 @@ export type BackofficeSection =
   | 'home'
   | 'catalog-seed'
   | 'health'
+  | 'openai-usage'
   | 'queue'
   | 'repair-batches'
   | 'reviews'
