+++ apps/backoffice/src/components/recommendation-evals/list.tsx
@@ -11,18 +11,7 @@ import {
 } from './providerUsageViewModel';
 import { buildRecommendationEvalPageHref, RecommendationEvalStatusBadge } from './shared';
 
-import type { BackofficeOpenAIUsageState, OpenAIUsagePeriodKey } from '../../lib/openAIUsage';
 import type { RecommendationEvalRun, RecommendationEvalRunPage } from '@pop-choice/shared';
-
-const USAGE_PERIOD_LABELS: Record<OpenAIUsagePeriodKey, string> = {
-  '24h': '24h',
-  '7d': '7d',
-  '30d': '30d',
-};
-
-function formatNumber(value: number): string {
-  return new Intl.NumberFormat('en-US').format(value);
-}
 
 function runOpenAIUsage(run: RecommendationEvalRun): { cost: string; usage: string } {
   const usage = providerUsageViewFromReport(run.report);
@@ -31,96 +20,6 @@ function runOpenAIUsage(run: RecommendationEvalRun): { cost: string; usage: stri
     cost: formatProviderUsageCost(usage?.cost ?? null),
     usage: providerUsageText(usage),
   };
-}
-
-function OpenAIUsagePeriodLinks({ active }: { active: OpenAIUsagePeriodKey }) {
-  return (
-    <div className="segmented-links" aria-label="OpenAI usage period">
-      {(Object.keys(USAGE_PERIOD_LABELS) as OpenAIUsagePeriodKey[]).map((period) => (
-        <a
-          aria-current={period === active ? 'page' : undefined}
-          className={period === active ? 'active' : undefined}
-          href={`/recommendation-evals?usagePeriod=${period}`}
-          key={period}
-        >
-          {USAGE_PERIOD_LABELS[period]}
-        </a>
-      ))}
-    </div>
-  );
-}
-
-function OpenAIUsageOverview({ state }: { state: BackofficeOpenAIUsageState }) {
-  if (state.status !== 'available') {
-    return (
-      <section className="panel openai-usage-panel">
-        <div className="panel-header">
-          <div>
-            <h2>OpenAI usage</h2>
-            <p className="panel-subtitle">Admin usage and cost telemetry</p>
-          </div>
-          <OpenAIUsagePeriodLinks active={state.period} />
-        </div>
-        <p className="empty">{state.message}</p>
-      </section>
-    );
-  }
-
-  const { summary } = state;
-  const categories = Object.entries(summary.usage.byCategory).filter(
-    ([, totals]) => totals.requests > 0 || totals.inputTokens > 0 || totals.outputTokens > 0,
-  );
-
-  return (
-    <section className="panel openai-usage-panel">
-      <div className="panel-header">
-        <div>
-          <h2>OpenAI usage</h2>
-          <p className="panel-subtitle">
-            {formatBackofficeDateTime(summary.period.startTime)} -{' '}
-            {formatBackofficeDateTime(summary.period.endTime)}
-          </p>
-        </div>
-        <OpenAIUsagePeriodLinks active={state.period} />
-      </div>
-      <div className="openai-usage-summary">
-        <div>
-          <span>Total cost</span>
-          <strong>{formatProviderUsageCost(summary.costs.total)}</strong>
-        </div>
-        <div>
-          <span>Requests</span>
-          <strong>{formatNumber(summary.usage.total.requests)}</strong>
-        </div>
-        <div>
-          <span>Input tokens</span>
-          <strong>{formatNumber(summary.usage.total.inputTokens)}</strong>
-        </div>
-        <div>
-          <span>Output tokens</span>
-          <strong>{formatNumber(summary.usage.total.outputTokens)}</strong>
-        </div>
-      </div>
-      {categories.length > 0 ? (
-        <DataTable
-          className="openai-usage-table"
-          columns={['Category', 'Requests', 'Input tokens', 'Output tokens', 'Cached input']}
-        >
-          {categories.map(([category, totals]) => (
-            <tr key={category}>
-              <td>{category}</td>
-              <td>{formatNumber(totals.requests)}</td>
-              <td>{formatNumber(totals.inputTokens)}</td>
-              <td>{formatNumber(totals.outputTokens)}</td>
-              <td>{formatNumber(totals.cachedInputTokens)}</td>
-            </tr>
-          ))}
-        </DataTable>
-      ) : (
-        <p className="empty">No OpenAI usage was reported for this period.</p>
-      )}
-    </section>
-  );
 }
 
 function RecommendationEvalRows({ runs }: { runs: RecommendationEvalRun[] }) {
@@ -222,11 +121,9 @@ function RecommendationEvalFlash({ status }: { status: string | null }) {
 }
 
 export function RecommendationEvalListPage({
-  openAIUsage,
   runPage,
   status,
 }: {
-  openAIUsage: BackofficeOpenAIUsageState;
   runPage: RecommendationEvalRunPage;
   status: string | null;
 }) {
@@ -239,6 +136,9 @@ export function RecommendationEvalListPage({
       actions={
         <>
           <ButtonLink href="/recommendation-evals">Refresh</ButtonLink>
+          <ButtonLink variant="quiet" href="/openai-usage">
+            OpenAI usage
+          </ButtonLink>
           <ButtonLink variant="quiet" href="/api/recommendation-evals">
             JSON
           </ButtonLink>
@@ -246,7 +146,6 @@ export function RecommendationEvalListPage({
       }
     >
       <RecommendationEvalFlash status={status} />
-      <OpenAIUsageOverview state={openAIUsage} />
       <section className="eval-run-panel" aria-labelledby="recommendation-eval-run-title">
         <div className="eval-run-panel-heading">
           <h2 id="recommendation-eval-run-title">Run eval</h2>
@@ -274,7 +173,6 @@ export function RecommendationEvalListPage({
                 buildRecommendationEvalPageHref({
                   page,
                   pageSize: runPage.limit,
-                  usagePeriod: openAIUsage.period,
                 })
               }
             />
@@ -307,7 +205,6 @@ export function RecommendationEvalListPage({
                 buildRecommendationEvalPageHref({
                   page,
                   pageSize: runPage.limit,
-                  usagePeriod: openAIUsage.period,
                 })
               }
             />
